feat: split archive api to a different file to simplify documentation
